Revert "port Automate vgpu vm upgrade test from cnv-tests repo (#982) (#1105)"
This reverts commit 76546987865344afdd838ee2ed4764b07a9772c0.
