+++ tests/virt/conftest.py
@@ -1,6 +1,4 @@
 import logging
-from multiprocessing import Manager, cpu_count
-from multiprocessing.dummy import Pool
 
 import pytest
 from bitmath import parse_string_unsafe
@@ -9,17 +7,8 @@ from ocp_resources.performance_profile import PerformanceProfile
 from ocp_resources.storage_profile import StorageProfile
 from pytest_testconfig import py_config
 
-from tests.virt.node.gpu.constants import (
-    GPU_CARDS_MAP,
-    NVIDIA_VGPU_MANAGER_DS,
-)
-from tests.virt.node.gpu.utils import (
-    wait_for_manager_pods_deployed,
-)
-from tests.virt.utils import check_node_for_missing_mdev_bus, patch_hco_cr_with_mdev_permitted_hostdevices
 from utilities.constants import AMD, INTEL
-from utilities.exceptions import UnsupportedGPUDeviceError
-from utilities.infra import exit_pytest_execution, label_nodes
+from utilities.infra import exit_pytest_execution
 from utilities.virt import get_nodes_gpu_info
 
 LOGGER = logging.getLogger(__name__)
@@ -168,66 +157,3 @@ def vm_cpu_flags(nodes_cpu_virt_extension):
         if nodes_cpu_virt_extension
         else None
     )
-
-
-@pytest.fixture(scope="session")
-def supported_gpu_device(workers_utility_pods, nodes_with_supported_gpus):
-    gpu_info = get_nodes_gpu_info(util_pods=workers_utility_pods, node=nodes_with_supported_gpus[0])
-    for gpu_id in GPU_CARDS_MAP:
-        if gpu_id in gpu_info:
-            return GPU_CARDS_MAP[gpu_id]
-
-    raise UnsupportedGPUDeviceError("GPU device ID not in current GPU_CARDS_MAP!")
-
-
-@pytest.fixture(scope="session")
-def hco_cr_with_mdev_permitted_hostdevices_scope_session(hyperconverged_resource_scope_session, supported_gpu_device):
-    yield from patch_hco_cr_with_mdev_permitted_hostdevices(
-        hyperconverged_resource=hyperconverged_resource_scope_session, supported_gpu_device=supported_gpu_device
-    )
-
-
-@pytest.fixture(scope="session")
-def gpu_nodes_labeled_with_vm_vgpu(nodes_with_supported_gpus):
-    yield from label_nodes(nodes=nodes_with_supported_gpus, labels={"nvidia.com/gpu.workload.config": "vm-vgpu"})
-
-
-@pytest.fixture(scope="session")
-def vgpu_ready_nodes(admin_client, gpu_nodes_labeled_with_vm_vgpu):
-    wait_for_manager_pods_deployed(admin_client=admin_client, ds_name=NVIDIA_VGPU_MANAGER_DS)
-    yield gpu_nodes_labeled_with_vm_vgpu
-
-
-@pytest.fixture(scope="session")
-def non_existent_mdev_bus_nodes(workers_utility_pods, vgpu_ready_nodes):
-    """
-    Check if the mdev_bus needed for vGPU is available.
-
-    On the Worker Node on which GPU Device exists, check if the
-    mdev_bus needed for vGPU is available.
-    If it's not available, this means the nvidia-vgpu-manager-daemonset
-    Pod might not be in running state in the nvidia-gpu-operator namespace.
-
-    Raises:
-        pytest.fail: If mdev_bus is missing on any nodes or if multiprocessing fails.
-    """
-    manager = Manager()
-    non_existent_mdev_bus_nodes = manager.list()
-
-    try:
-        with Pool(processes=min(len(vgpu_ready_nodes), cpu_count())) as pool:
-            pool.map(
-                check_node_for_missing_mdev_bus,
-                [(node, non_existent_mdev_bus_nodes, workers_utility_pods) for node in vgpu_ready_nodes],
-            )
-    except Exception as e:
-        pytest.fail(f"Failed to check mdev_bus on nodes: {e}")
-
-    if non_existent_mdev_bus_nodes:
-        pytest.fail(
-            reason=(
-                f"On these nodes: {list(non_existent_mdev_bus_nodes)} "
-                "mdev_bus is not available. Ensure that in 'nvidia-gpu-operator' "
-                "namespace the nvidia-vgpu-manager-daemonset Pod is Running."
-            )
-        )
